Move logic to create ProjectData into ProjectData class
